+++ docs/CHANGE_LOG.md
@@ -10,6 +10,8 @@ Change the contents of this file to this:
 - Added functionality to cater for new Internal Alert in Data Store.
 - Added Tests for new Internal Alerts in System/Github Alerter and Alert Store.
 - Added the ChainlinkNodeMonitor, NodeMonitorsManager, and their tests.
+- Refactored RabbitMQ queues and routing keys.
+- The SystemMonitorsManager additionally now parses systems belonging to chains from the `system_config.ini` if Chainlink is the base chain. Same schema as `GENERAL` is expected.
 
 ## 0.1.2